requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
stripe>=9.0.0
pandas>=2.2.0
numpy>=1.26.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
